+++ src/index.js
@@ -13,5 +13,5 @@
  * @namespace Tiny.particles
  */
 
-export {default as ParticleContainer} from './ParticleContainer';
-export {default as ParticleRenderer} from './webgl/ParticleRenderer';
+export { default as ParticleContainer } from './ParticleContainer';
+export { default as ParticleRenderer } from './webgl/ParticleRenderer';