+++ src/webgl/ParticleBuffer.js
@@ -3,10 +3,8 @@ import createIndicesForQuads from '../utils/createIndicesForQuads';
 /**
  * @author Mat Groves
  *
- * Big thanks to the very clever Matt DesLauriers <mattdesl> https://github.com/mattdesl/
- * for creating the original pixi version!
- * Also a thanks to https://github.com/bchevalier for tweaking the tint and alpha so that
- * they now share 4 bytes on the vertex buffer
+ * Big thanks to the very clever Matt DesLauriers <mattdesl> https://github.com/mattdesl/ for creating the original pixi version!
+ * Also a thanks to https://github.com/bchevalier for tweaking the tint and alpha so that they now share 4 bytes on the vertex buffer
  *
  * Heavily inspired by LibGDX's ParticleBuffer:
  * https://github.com/libgdx/libgdx/blob/master/gdx/src/com/badlogic/gdx/graphics/g2d/ParticleBuffer.java
@@ -32,20 +30,6 @@ class ParticleBuffer {
      * @member {WebGLRenderingContext}
      */
     this.gl = gl;
-
-    /**
-     * Size of a single vertex.
-     *
-     * @member {number}
-     */
-    this.vertSize = 2;
-
-    /**
-     * Size of a single vertex in bytes.
-     *
-     * @member {number}
-     */
-    this.vertByteSize = this.vertSize * 4;
 
     /**
      * The number of particles the buffer can hold
@@ -77,6 +61,7 @@ class ParticleBuffer {
         attribute: property.attribute,
         size: property.size,
         uploadFunction: property.uploadFunction,
+        unsignedByte: property.unsignedByte,
         offset: property.offset,
       };
 
@@ -90,10 +75,14 @@ class ParticleBuffer {
     this.staticStride = 0;
     this.staticBuffer = null;
     this.staticData = null;
+    this.staticDataUint32 = null;
 
     this.dynamicStride = 0;
     this.dynamicBuffer = null;
     this.dynamicData = null;
+    this.dynamicDataUint32 = null;
+
+    this._updateID = 0;
 
     this.initBuffers();
   }
@@ -125,8 +114,11 @@ class ParticleBuffer {
       this.dynamicStride += property.size;
     }
 
-    this.dynamicData = new Float32Array(this.size * this.dynamicStride * 4);
-    this.dynamicBuffer = Tiny.glCore.GLBuffer.createVertexBuffer(gl, this.dynamicData, gl.STREAM_DRAW);
+    const dynBuffer = new ArrayBuffer(this.size * this.dynamicStride * 4 * 4);
+
+    this.dynamicData = new Float32Array(dynBuffer);
+    this.dynamicDataUint32 = new Uint32Array(dynBuffer);
+    this.dynamicBuffer = Tiny.glCore.GLBuffer.createVertexBuffer(gl, dynBuffer, gl.STREAM_DRAW);
 
     // static //
     let staticOffset = 0;
@@ -141,8 +133,11 @@ class ParticleBuffer {
       this.staticStride += property.size;
     }
 
-    this.staticData = new Float32Array(this.size * this.staticStride * 4);
-    this.staticBuffer = Tiny.glCore.GLBuffer.createVertexBuffer(gl, this.staticData, gl.STATIC_DRAW);
+    const statBuffer = new ArrayBuffer(this.size * this.staticStride * 4 * 4);
+
+    this.staticData = new Float32Array(statBuffer);
+    this.staticDataUint32 = new Uint32Array(statBuffer);
+    this.staticBuffer = Tiny.glCore.GLBuffer.createVertexBuffer(gl, statBuffer, gl.STATIC_DRAW);
 
     this.vao = new Tiny.glCore.VertexArrayObject(gl)
       .addIndex(this.indexBuffer);
@@ -150,27 +145,49 @@ class ParticleBuffer {
     for (let i = 0; i < this.dynamicProperties.length; ++i) {
       const property = this.dynamicProperties[i];
 
-      this.vao.addAttribute(
-        this.dynamicBuffer,
-        property.attribute,
-        gl.FLOAT,
-        false,
-        this.dynamicStride * 4,
-        property.offset * 4
-      );
+      if (property.unsignedByte) {
+        this.vao.addAttribute(
+          this.dynamicBuffer,
+          property.attribute,
+          gl.UNSIGNED_BYTE,
+          true,
+          this.dynamicStride * 4,
+          property.offset * 4
+        );
+      } else {
+        this.vao.addAttribute(
+          this.dynamicBuffer,
+          property.attribute,
+          gl.FLOAT,
+          false,
+          this.dynamicStride * 4,
+          property.offset * 4
+        );
+      }
     }
 
     for (let i = 0; i < this.staticProperties.length; ++i) {
       const property = this.staticProperties[i];
 
-      this.vao.addAttribute(
-        this.staticBuffer,
-        property.attribute,
-        gl.FLOAT,
-        false,
-        this.staticStride * 4,
-        property.offset * 4
-      );
+      if (property.unsignedByte) {
+        this.vao.addAttribute(
+          this.staticBuffer,
+          property.attribute,
+          gl.UNSIGNED_BYTE,
+          true,
+          this.staticStride * 4,
+          property.offset * 4
+        );
+      } else {
+        this.vao.addAttribute(
+          this.staticBuffer,
+          property.attribute,
+          gl.FLOAT,
+          false,
+          this.staticStride * 4,
+          property.offset * 4
+        );
+      }
     }
   }
 
@@ -185,7 +202,9 @@ class ParticleBuffer {
     for (let i = 0; i < this.dynamicProperties.length; i++) {
       const property = this.dynamicProperties[i];
 
-      property.uploadFunction(children, startIndex, amount, this.dynamicData, this.dynamicStride, property.offset);
+      property.uploadFunction(children, startIndex, amount,
+        property.unsignedByte ? this.dynamicDataUint32 : this.dynamicData,
+        this.dynamicStride, property.offset);
     }
 
     this.dynamicBuffer.upload();
@@ -202,7 +221,9 @@ class ParticleBuffer {
     for (let i = 0; i < this.staticProperties.length; i++) {
       const property = this.staticProperties[i];
 
-      property.uploadFunction(children, startIndex, amount, this.staticData, this.staticStride, property.offset);
+      property.uploadFunction(children, startIndex, amount,
+        property.unsignedByte ? this.staticDataUint32 : this.staticData,
+        this.staticStride, property.offset);
     }
 
     this.staticBuffer.upload();
@@ -214,12 +235,16 @@ class ParticleBuffer {
    */
   destroy() {
     this.dynamicProperties = null;
-    this.dynamicData = null;
     this.dynamicBuffer.destroy();
+    this.dynamicBuffer = null;
+    this.dynamicData = null;
+    this.dynamicDataUint32 = null;
 
     this.staticProperties = null;
-    this.staticData = null;
     this.staticBuffer.destroy();
+    this.staticBuffer = null;
+    this.staticData = null;
+    this.staticDataUint32 = null;
   }
 }