+++ src/webgl/ParticleShader.js
@@ -11,48 +11,41 @@ class ParticleShader extends Tiny.Shader {
     super(
       gl,
       // vertex shader
-      [
-        'attribute vec2 aVertexPosition;',
-        'attribute vec2 aTextureCoord;',
-        'attribute float aColor;',
+      `attribute vec2 aVertexPosition;
+attribute vec2 aTextureCoord;
+attribute vec4 aColor;
 
-        'attribute vec2 aPositionCoord;',
-        'attribute vec2 aScale;',
-        'attribute float aRotation;',
+attribute vec2 aPositionCoord;
+attribute float aRotation;
 
-        'uniform mat3 projectionMatrix;',
+uniform mat3 projectionMatrix;
+uniform vec4 uColor;
 
-        'varying vec2 vTextureCoord;',
-        'varying float vColor;',
+varying vec2 vTextureCoord;
+varying vec4 vColor;
 
-        'void main(void){',
-        '   vec2 v = aVertexPosition;',
+void main(void){
+    float x = (aVertexPosition.x) * cos(aRotation) - (aVertexPosition.y) * sin(aRotation);
+    float y = (aVertexPosition.x) * sin(aRotation) + (aVertexPosition.y) * cos(aRotation);
 
-        '   v.x = (aVertexPosition.x) * cos(aRotation) - (aVertexPosition.y) * sin(aRotation);',
-        '   v.y = (aVertexPosition.x) * sin(aRotation) + (aVertexPosition.y) * cos(aRotation);',
-        '   v = v + aPositionCoord;',
+    vec2 v = vec2(x, y);
+    v = v + aPositionCoord;
 
-        '   gl_Position = vec4((projectionMatrix * vec3(v, 1.0)).xy, 0.0, 1.0);',
+    gl_Position = vec4((projectionMatrix * vec3(v, 1.0)).xy, 0.0, 1.0);
 
-        '   vTextureCoord = aTextureCoord;',
-        '   vColor = aColor;',
-        '}',
-      ].join('\n'),
+    vTextureCoord = aTextureCoord;
+    vColor = aColor * uColor;
+}`,
       // hello
-      [
-        'varying vec2 vTextureCoord;',
-        'varying float vColor;',
+      `varying vec2 vTextureCoord;
+varying vec4 vColor;
 
-        'uniform sampler2D uSampler;',
-        'uniform float uAlpha;',
-        'uniform vec3 tint;',
+uniform sampler2D uSampler;
 
-        'void main(void){',
-        '  vec4 color = texture2D(uSampler, vTextureCoord) * vColor * vec4(tint * uAlpha, uAlpha);',
-        '  if (color.a == 0.0) discard;',
-        '  gl_FragColor = color;',
-        '}',
-      ].join('\n')
+void main(void){
+  vec4 color = texture2D(uSampler, vTextureCoord) * vColor;
+  gl_FragColor = color;
+}`
     );
   }
 }