+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.1.0
+
+`2018-11-30`
+- 此版本需要基于 TinyJS v1.2.0
+- 构建方式从 webpack@1.* 换成 rollup
+
+### Added
+- 可以支持 `tint` 属性
+- 增加了 `autoResize` 参数
+
+### Fixed
+- `ParticleBuffer.destroy()` 时清理了一些属性
+- 支持多重渲染
+
 ## 0.0.3
 
 `2017-12-12`